# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "ihavefood"
  description: "Your API description"
  version: 0.1.0
servers:
    - url: "http://localhost:180"
      description: "abcdefg"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      description: Register a new user 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "yuki002"
                email:
                  type: string
                  format: email
                  example: "yuki_sleep@mail.com"
                password:
                  type: string
                  example: "secret"
                phone_number:
                  type: string
                  example: "081 222 3333"
                address:
                  type: object
                  properties:
                    address_name:
                      type: string
                      example: "Jai Dee Hotel 123 Moo 4"
                    sub_district:
                      type: string
                      example: "Nimmanhaemin"
                    district:
                      type: string
                      example: "Mueang"
                    province:
                      type: string
                      example: "Chiang Mai"
                    postal_code:
                      type: string
                      example: "50200"
              required:
                - username
                - email
                - password
                - phone_number
                - address
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
  /auth/login:
    post:
      tags:
        - auth
      summary: Logs user into the system
      description: "XXXXXXXXXXX"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  default: yuki002
                password:
                  type: string
                  default: secret
      responses:
        '200':
          description: Successful login with access token and token expiration time (Unix)

#---------------User Service-----------------
...
  /api/orders/place-order:
    post:
      tags:
      - order 
      summary: Place an order
      description: Endpoint to place an order at a restaurant.
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: yuki002
                restaurant_name:
                  type: string
                  example: HaiDee
                menus:
                  type: array
                  items:
                    type: object
                    properties:
                      food_name:
                        type: string
                        example: Pad Thai
                      price:
                        type: number
                        example: 50
                delivery_fee:
                  type: number
                  example: 50
                coupon_code:
                  type: string
                  example: 129dh012
                coupon_discount:
                  type: number
                  example: 20
                total:
                  type: number
                  example: 150
                address:
                  type: object
                  properties:
                    address_name:
                      type: string
                      example: home
                    sub_district:
                      type: string
                      example: Nong Hoi
                    district:
                      type: string
                      example: Mueang Chiang Mai
                    province:
                      type: string
                      example: Chiang Mai
                    postalCode:
                      type: string
                      example: 50000
                    country:
                      type: string
                      example: Thailand
                contact:
                  type: object
                  properties:
                    phone_number:
                      type: string
                      example: +668 1234 5678
                    email:
                      type: string
                      example: r7do@mail.com
                payment_method:
                  type: integer
                  description: |
                    Payment method used for the order.
                    0 = Cash
                    1 = Credit Card
                  example: 0
              required:
                - username
                - restaurant_name
                - menus
                - delivery_fee
                - total
                - address
                - contact
                - payment_method
      responses:
        '200':
          description: Successfully placed an order
        '400':
          description: Invalid request
      security:
        - BearerAuth: [] 
...